Update User Repository and NextAuth configuration
