+++ src/api/lib/infrastructure/persistent/mongo/UserRepository.ts
@@ -8,7 +8,8 @@ export class UserRepository implements IUserRepository {
   private readonly MongooseUserModel;
 
   constructor() {
-    this.MongooseUserModel = mongoose.model("User", UserSchema);
+    this.MongooseUserModel =
+      mongoose.models.User || mongoose.model("User", UserSchema);
   }
 
   async createUser(user: User): Promise<void> {
