+++ src/provider/NextAuthProvider.tsx
@@ -1,0 +1,12 @@
+"use client";
+import { SessionProvider } from "next-auth/react";
+import { Session } from "next-auth";
+export const NextAuthProvider = ({
+  children,
+  session,
+}: {
+  children: React.ReactNode;
+  session: Session | null;
+}) => {
+  return <SessionProvider session={session}>{children}</SessionProvider>;
+};
